Finishing setup our app

+++ backend/src/controllers/notesController.ts
@@ -2,11 +2,16 @@ import { RequestHandler } from "express";
 import NoteModel from '../moduls/note'
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { DefineId } from "../Util/DefineId";
 
 export const getNotes: RequestHandler = async  (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
 
   try {
-    const notes = await NoteModel.find().exec();
+
+    DefineId(authenticatedUserId);
+    
+    const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -15,13 +20,22 @@ export const getNotes: RequestHandler = async  (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+
+    DefineId(authenticatedUserId);
+
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if(!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "You are not allowed to get this note");
     }
     res.status(200).json(note);
 
@@ -36,13 +50,15 @@ interface CreateNotesBody {
 }
 
 export  const createNotes: RequestHandler<unknown, unknown, CreateNotesBody, unknown> = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const title = req.body.title;
   const text = req.body.text;
   try {
+    DefineId(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must have a title!")
     }
-    const newNote = await NoteModel.create({ title, text });
+    const newNote = await NoteModel.create({ title, text, userId: authenticatedUserId });
 
     res.status(201).json(newNote);
   } catch (error) {
@@ -65,8 +81,11 @@ export const updateNote: RequestHandler<updateNoteParams, unknown,updateNoteBody
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
   
   try {
+
+    DefineId(authenticatedUserId);
 
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
@@ -81,6 +100,10 @@ export const updateNote: RequestHandler<updateNoteParams, unknown,updateNoteBody
       throw createHttpError(404, "Note not found");
     }
 
+    if(!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "You cannot access this note");
+    }
+
     note.title = newTitle;
     note.text = newText;
 
@@ -93,8 +116,10 @@ export const updateNote: RequestHandler<updateNoteParams, unknown,updateNoteBody
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    DefineId(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -104,6 +129,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
       throw createHttpError(404, "Note not found");
     }
     
+    if(!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "You cannot delete this note");
+    }
+
     await note.deleteOne();
     res.sendStatus(204);
